refactor(data-store): bryt ut LocalStore-kärna + IndexedDB-persistens (#412)
Första kod-steget i ADR 0016-migreringen. Den generiska in-memory-motorn
(delegates, snapshot/rollback-transaktioner, write-back-hook) bryts ut ur
DemoDataStore till en återanvändbar `LocalStore` — kärnan som både demon och
den framtida offline-cachen bygger på. DemoDataStore blir en tunn subklass
(oförändrat beteende; 110 befintliga data-store/router-tester gröna).

Lägger till persistens-porten `LocalStorePersistence` (hydrera/spara hela
källan) med två adaptrar: `InMemoryPersistence` (tester/demo) och
`IndexedDbPersistence` (browser, injicerbar IDBFactory → testbar via
fake-indexeddb). `createPersistedLocalStore` hydrerar källan och skriver
tillbaka efter varje mutation. Finkornig mutations-kö/delta-sync kommer i
#413/#414 (ADR 0017).

Tester: LocalStore-kärna (read-only/writable/tx-rollback/currentSource),
InMemoryPersistence + factory (hydrate/seed/write-back), IndexedDbPersistence
(round-trip + Date-bevaring) mot fake-indexeddb. Coverage-golv hålls (87.69%).

Closes #412
