+++ package.json
@@ -108,6 +108,7 @@
     "eslint": "^10",
     "eslint-config-next": "16.2.7",
     "eslint-plugin-import-x": "^4.16.2",
+    "fake-indexeddb": "^6.2.5",
     "husky": "^9.1.7",
     "jscpd": "^5.0.4",
     "knip": "^6.16.1",
